feat: bonus objects parsers and builders added and functional
