+++ libs/librt/include/rtelm.h
@@ -6,7 +6,7 @@
 /*   By: sscheini <sscheini@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/19 17:20:30 by sscheini          #+#    #+#             */
-/*   Updated: 2026/05/17 15:13:20 by sscheini         ###   ########.fr       */
+/*   Updated: 2026/05/19 20:19:57 by sscheini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,14 +16,12 @@
 # include "rtmlx.h"
 # include "rtmth.h"
 # include "libft.h"
-
-/*--------------------------------------------------------------------------*/
-/*--------------------------------STRUCTURES--------------------------------*/
-/*--------------------------------------------------------------------------*/
-
 # if BONUS
 #  include "rtelm_bonus.h"
 # else
+/*--------------------------------------------------------------------------*/
+/*--------------------------------STRUCTURES--------------------------------*/
+/*--------------------------------------------------------------------------*/
 
 /**
  * Represents a point light source in the scene.
