+++ libs/librt/include/rterr.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   rterr.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aramos-r <aramos-r@student.42malaga.com    +#+  +:+       +#+        */
+/*   By: sscheini <sscheini@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/19 17:19:49 by sscheini          #+#    #+#             */
-/*   Updated: 2026/04/23 19:41:55 by aramos-r         ###   ########.fr       */
+/*   Updated: 2026/05/19 20:15:34 by sscheini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,9 +19,6 @@
 # include <sys/wait.h>
 # include <sys/time.h>
 # include <sys/types.h>
-# if BONUS
-#  include "rterr_bonus.h"
-# endif
 
 /*--------------------------------------------------------------------------*/
 /*----------------------------------MEMFREE---------------------------------*/
