+++ libs/librt/src/rtelm/camera.c
@@ -6,7 +6,7 @@
 /*   By: sscheini <sscheini@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/25 21:14:30 by sscheini          #+#    #+#             */
-/*   Updated: 2026/04/25 17:20:40 by sscheini         ###   ########.fr       */
+/*   Updated: 2026/05/19 20:20:10 by sscheini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
