+++ libs/librt/src/rtelm/lights_bonus.c
@@ -6,7 +6,7 @@
 /*   By: sscheini <sscheini@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/12 15:47:37 by sscheini          #+#    #+#             */
-/*   Updated: 2026/05/19 19:58:14 by sscheini         ###   ########.fr       */
+/*   Updated: 2026/05/19 20:59:14 by sscheini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
