+++ src/rtapp/init/parser/parse_objects_bonus.c
@@ -6,13 +6,13 @@
 /*   By: sscheini <sscheini@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/17 17:31:17 by sscheini          #+#    #+#             */
-/*   Updated: 2026/05/19 19:53:22 by sscheini         ###   ########.fr       */
+/*   Updated: 2026/05/19 20:40:59 by sscheini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "rtapp_parser.h"
 
-static int	parse_arg(char **arr, int i, const char *const msgs[], int e)
+static int	parse_arg(char **arr, int i, const char *const *msgs, int e)
 {
 	const char	*err = "[line: %i][%s] parser failed: %s";
 	int			j;
@@ -26,11 +26,12 @@ static int	parse_arg(char **arr, int i, const char *const msgs[], int e)
 	return (RT_SUCCESS);
 }
 
-int	parse_quadric(char **arr, int i, static const char *const *g_msgs)
+int	parse_quadric(char **arr, int i, const char *const *g_msgs)
 {
 	const char	*err = "[line: %i][%s] parser failed: %s";
+	int			arrlen;
 
-	if (parse_arg(arr, i, (*g_msgs), 5))
+	if (parse_arg(arr, i, g_msgs, 5))
 		return (RT_FAILURE);
 	if (parse_vector(arr[0], arr[1], i, build_limit(-FLT_MAX, FLT_MAX)))
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid coordinates."));
@@ -42,19 +43,23 @@ int	parse_quadric(char **arr, int i, static const char *const *g_msgs)
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid height."));
 	if (parse_vector(arr[0], arr[5], i, build_limit(0, 255)))
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid color."));
-	if (arr[6] && parse_double(arr[0], arr[6], i, build_limit(0, FLT_MAX)))
-		if (parse_extention(arr[0], arr[6], i, ".map"))
+	arrlen = ft_arglen(arr);
+	if (arrlen > 4 && arr[4]
+		&& parse_double(arr[0], arr[4], i, build_limit(0, FLT_MAX)))
+		if (parse_extention(arr[0], arr[4], i, ".map"))
 			return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid bonus attributes."));
-	if (arr[7] && parse_extention(arr[0], arr[6], i, ".map"))
+	if (arrlen > 5 && arr[5]
+		&& parse_extention(arr[0], arr[5], i, ".map"))
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid bonus attributes."));
 	return (RT_SUCCESS);
 }
 
-int	parse_sp(char **arr, int i, static const char *const *g_msgs)
+int	parse_sp(char **arr, int i, const char *const *g_msgs)
 {
 	const char	*err = "[line: %i][%s] parser failed: %s";
+	int			arrlen;
 
-	if (parse_arg(arr, i, g_sp_msgs, 3))
+	if (parse_arg(arr, i, g_msgs, 3))
 		return (RT_FAILURE);
 	if (parse_vector(arr[0], arr[1], i, build_limit(-FLT_MAX, FLT_MAX)))
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid coordinates."));
@@ -62,19 +67,23 @@ int	parse_sp(char **arr, int i, static const char *const *g_msgs)
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid radius."));
 	if (parse_vector(arr[0], arr[3], i, build_limit(0, 255)))
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid color."));
-	if (arr[4] && parse_double(arr[0], arr[4], i, build_limit(0, FLT_MAX)))
+	arrlen = ft_arglen(arr);
+	if (arrlen > 4 && arr[4]
+		&& parse_double(arr[0], arr[4], i, build_limit(0, FLT_MAX)))
 		if (parse_extention(arr[0], arr[4], i, ".map"))
 			return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid bonus attributes."));
-	if (arr[5] && parse_extention(arr[0], arr[5], i, ".map"))
+	if (arrlen > 5 && arr[5]
+		&& parse_extention(arr[0], arr[5], i, ".map"))
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid bonus attributes."));
 	return (RT_SUCCESS);
 }
 
-int	parse_pl(char **arr, int i, static const char *const *g_msgs)
+int	parse_pl(char **arr, int i, const char *const *g_msgs)
 {
 	const char	*err = "[line: %i][%s] parser failed: %s";
+	int			arrlen;
 
-	if (parse_arg(arr, i, g_pl_msgs, 3))
+	if (parse_arg(arr, i, g_msgs, 3))
 		return (RT_FAILURE);
 	if (parse_vector(arr[0], arr[1], i, build_limit(-FLT_MAX, FLT_MAX)))
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid coordinates."));
@@ -82,10 +91,13 @@ int	parse_pl(char **arr, int i, static const char *const *g_msgs)
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid normal."));
 	if (parse_vector(arr[0], arr[3], i, build_limit(0, 255)))
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid color."));
-	if (arr[4] && parse_double(arr[0], arr[4], i, build_limit(0, FLT_MAX)))
+	arrlen = ft_arglen(arr);
+	if (arrlen > 4 && arr[4]
+		&& parse_double(arr[0], arr[4], i, build_limit(0, FLT_MAX)))
 		if (parse_extention(arr[0], arr[4], i, ".map"))
 			return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid bonus attributes."));
-	if (arr[5] && parse_extention(arr[0], arr[5], i, ".map"))
+	if (arrlen > 5 && arr[5]
+		&& parse_extention(arr[0], arr[5], i, ".map"))
 		return (rtlog(RT_ERRLOG, 0, err, i, arr[0], "invalid bonus attributes."));
 	return (RT_SUCCESS);
 }
