+++ tests_bonus/combinationAll.rt
@@ -1,7 +1,11 @@
 A 0.2 255,255,255
 C 0,15,-40 0,-0.2,1 70
-L 0,30,-10 0.9 255,255,255
+L 0,30,-10 0.9 0,255,0
 
 pl 0,-10,0 0,1,0 100,100,100
 sp -15,0,30 8 255,0,0
 cy 15,0,30 0,1,0 8 20 0,0,255
+
+hb -10,0,20 0,1,0 6 12 255,255,0
+pb 0,0,25 0,1,0 6 12 255,0,255
+cn 10,0,20 0,1,0 6 12 0,255,255
